+++ src/kernel/relu.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 
 #define BUFFER_SIZE  512
-__mram_noinit int32_t buffer[BUFFER_SIZE];
+__mram_noinit float buffer[BUFFER_SIZE];
 __mram_noinit int32_t params[4];
 
 
 int main() {
-    __dma_aligned int32_t local_cache[BUFFER_SIZE];
+    __dma_aligned float local_cache[BUFFER_SIZE];
 
     int actionable_length = params[0];
     printf("Actionable length: %d\n", actionable_length);
